Rename and move data models
Renaming and moving the data models
