+++ data/src/main/java/com/kennedydias/data/model/MovieRatingModel.kt
@@ -1,8 +1,8 @@
-package com.kennedydias.data.api.responsemodel
+package com.kennedydias.data.model
 
 import com.google.gson.annotations.SerializedName
 
-data class MovieRatingResponseModel(
+data class MovieRatingModel(
     @SerializedName("Source") val source: String? = null,
     @SerializedName("Value") val value: String? = null
 )
