+++ data/src/main/java/com/kennedydias/data/model/MoviesListModel.kt
@@ -1,9 +1,9 @@
-package com.kennedydias.data.api.responsemodel
+package com.kennedydias.data.model
 
 import com.google.gson.annotations.SerializedName
 
-data class MoviesListResponseModel(
-    @SerializedName("Search") val search: List<MovieShortResponseModel>? = null,
+data class MoviesListModel(
+    @SerializedName("Search") val search: List<MovieShortModel>? = null,
     @SerializedName("totalResults") val totalResults: String? = null,
     @SerializedName("Response") val response: String? = null,
     @SerializedName("Error") val error: String? = null
