+++ domain/src/main/java/com/kennedydias/domain/usecase/GetSeriesDetailsUseCase.kt
@@ -1,17 +1,17 @@
 package com.kennedydias.domain.usecase
 
-import com.kennedydias.data.api.repository.MoviesRepository
-import com.kennedydias.data.api.responsemodel.MovieFullResponseModel
+import com.kennedydias.data.model.MovieFullModel
+import com.kennedydias.data.remote.repository.MoviesRepository
 
 class GetSeriesDetailsUseCase(
     private val moviesRepository: MoviesRepository
-) : UseCase<MovieFullResponseModel>() {
+) : UseCase<MovieFullModel>() {
 
     var imdbId: String? = null
     var title: String? = null
     var year: String? = null
 
-    override suspend fun executeOnBackground(): MovieFullResponseModel {
+    override suspend fun executeOnBackground(): MovieFullModel {
         return moviesRepository.getMovieDetails(
             imdbId = imdbId,
             title = title,
