+++ domain/src/main/java/com/kennedydias/domain/usecase/GetSeriesListUseCase.kt
@@ -1,17 +1,17 @@
 package com.kennedydias.domain.usecase
 
-import com.kennedydias.data.api.repository.MoviesRepository
-import com.kennedydias.data.api.responsemodel.MoviesListResponseModel
+import com.kennedydias.data.model.MoviesListModel
+import com.kennedydias.data.remote.repository.MoviesRepository
 
 class GetSeriesListUseCase(
     private val moviesRepository: MoviesRepository
-) : UseCase<MoviesListResponseModel>() {
+) : UseCase<MoviesListModel>() {
 
     var search: String? = null
     var year: String? = null
     var page: Int? = null
 
-    override suspend fun executeOnBackground(): MoviesListResponseModel {
+    override suspend fun executeOnBackground(): MoviesListModel {
         return moviesRepository.getMoviesList(
             search = search,
             type = "series",
